v2/requirements.txt
openai
streamlit>=1.40
python-dotenv>=1.0
pandas>=2.2
